+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -1,7 +1,11 @@
-package nextstep.subway.line;
+package nextstep.subway.line.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import nextstep.subway.line.infra.LineRepository;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.Section;
+import nextstep.subway.line.domain.Sections;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.line.dto.SectionRequest;
@@ -9,8 +13,8 @@ import nextstep.subway.line.dto.SectionResponse;
 import nextstep.subway.line.dto.UpdateLineRequest;
 import nextstep.subway.line.exception.NoLineException;
 import nextstep.subway.line.exception.NoStationException;
-import nextstep.subway.station.Station;
-import nextstep.subway.station.StationRepository;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.station.infra.StationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
