+++ src/main/java/nextstep/subway/line/infra/LineRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.line;
+package nextstep.subway.line.infra;
 
+import nextstep.subway.line.domain.Line;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
