+++ src/main/java/nextstep/subway/station/dto/StationRequest.java
@@ -1,4 +1,6 @@
-package nextstep.subway.station;
+package nextstep.subway.station.dto;
+
+import nextstep.subway.station.domain.Station;
 
 public class StationRequest {
