+++ src/main/java/nextstep/subway/station/dto/StationResponse.java
@@ -1,6 +1,7 @@
-package nextstep.subway.station;
+package nextstep.subway.station.dto;
 
 import java.time.LocalDateTime;
+import nextstep.subway.station.domain.Station;
 
 public class StationResponse {
     private Long id;
