+++ src/main/java/nextstep/subway/station/infra/StationRepository.java
@@ -1,6 +1,7 @@
-package nextstep.subway.station;
+package nextstep.subway.station.infra;
 
 import java.util.List;
+import nextstep.subway.station.domain.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
